% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\name{game}
\alias{game}
\title{Play Game in Shiny App}
\usage{
game(type = "corona")
}
\arguments{
\item{type}{Either "corona" or "dino". Default : "corona"}
}
\value{

}
\description{
Play Game in Shiny App
}
\examples{
library(shiny)
library(shinyGames)
ui <- fluidPage( game(type = "corona") )
server <- function(input, output, session) { }
shinyApp(ui, server)
}
